postgres/requirements.txt
psycopg2-binary
pydantic
pytest
sqlalchemy
